db_update_service/requirements.txt
python-dateutil
mysql-connector-python
requests
